Autenticação por parâmetro hash para todas as requisições de controller

+++ src/main/java/com/gabriel/martins/apinaescola/controller/AlunoController.java
@@ -2,8 +2,10 @@ package com.gabriel.martins.apinaescola.controller;
 
 import com.gabriel.martins.apinaescola.model.entity.AlunoEntity;
 import com.gabriel.martins.apinaescola.model.entity.EscolaEntity;
+import com.gabriel.martins.apinaescola.model.entity.UsuarioEntity;
 import com.gabriel.martins.apinaescola.model.service.AlunoService;
 import com.gabriel.martins.apinaescola.model.service.EscolaService;
+import com.gabriel.martins.apinaescola.model.service.UsuarioService;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,7 +17,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -29,18 +30,26 @@ public class AlunoController {
     @Autowired
     private EscolaService escolaService;
     
+    @Autowired
+    private UsuarioService usuarioService;
+    
     @PostMapping("/aluno")
     public ResponseEntity<?> cadastrarAluno(@RequestParam("hash") String hash, @RequestBody AlunoEntity aluno){
         if(hash.isBlank() || hash.isEmpty()){
             try {
-                EscolaEntity escola = escolaService.findById(aluno.getEscola().getId());
-                String codigoGerado = service.gerarCodigoEscola(escola, aluno.getUsuario().getCpf());
+                UsuarioEntity user = usuarioService.findByHash(hash);
+                if(user != null){
+                    EscolaEntity escola = escolaService.findById(aluno.getEscola().getId());
+                    String codigoGerado = service.gerarCodigoEscola(escola, aluno.getUsuario().getCpf());
 
-                aluno.setEscola(escola);
-                aluno.setCodigoGeradoEscola(codigoGerado);
+                    aluno.setEscola(escola);
+                    aluno.setCodigoGeradoEscola(codigoGerado);
 
-                service.save(aluno);
-                return ResponseEntity.status(HttpStatus.ACCEPTED).body(aluno);
+                    service.save(aluno);
+                    return ResponseEntity.status(HttpStatus.ACCEPTED).body(aluno);
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
+                }
             } catch (Exception e) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Não foi possível cadastrar.");
             }
@@ -53,8 +62,13 @@ public class AlunoController {
     public ResponseEntity<?> listarAlunosEscola(@RequestParam("hash") String hash, @PathVariable Long id){
         if(hash.isBlank() || hash.isEmpty()){
             try {
-                List<AlunoEntity> alunos = service.findByEscola(id);
-                return ResponseEntity.status(HttpStatus.ACCEPTED).body(alunos);
+                UsuarioEntity user = usuarioService.findByHash(hash);
+                if(user != null){
+                    List<AlunoEntity> alunos = service.findByEscola(id);
+                    return ResponseEntity.status(HttpStatus.ACCEPTED).body(alunos);
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
+                }
             } catch (Exception e) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Não foi possível encontrar os alunos.");
             }
@@ -67,8 +81,13 @@ public class AlunoController {
     public ResponseEntity<?> buscarAlunoPorId(@RequestParam("hash") String hash, @PathVariable Long id){
         if(hash.isBlank() || hash.isEmpty()){
             try {
-                AlunoEntity aluno = service.findById(id);
-                return ResponseEntity.status(HttpStatus.ACCEPTED).body(aluno);
+                UsuarioEntity user = usuarioService.findByHash(hash);
+                if(user != null){
+                    AlunoEntity aluno = service.findById(id);
+                    return ResponseEntity.status(HttpStatus.ACCEPTED).body(aluno);
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
+                }
             } catch (Exception e) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Não foi possível encontrar o aluno.");
             }
@@ -81,8 +100,13 @@ public class AlunoController {
     public ResponseEntity<?> atualizarAlunoPorId(@RequestParam("hash") String hash, @PathVariable Long id, @RequestBody AlunoEntity aluno) {
         if(hash.isBlank() || hash.isEmpty()){
             try {
-                service.update(aluno);
-                return ResponseEntity.status(HttpStatus.ACCEPTED).body(aluno);
+                UsuarioEntity user = usuarioService.findByHash(hash);
+                if(user != null){
+                    service.update(aluno);
+                    return ResponseEntity.status(HttpStatus.ACCEPTED).body(aluno);
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
+                }
             } catch (Exception e) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Requisição inválida.");
             }
@@ -95,8 +119,13 @@ public class AlunoController {
     public ResponseEntity<?> removerAlunoPorId(@RequestParam("hash") String hash, @PathVariable Long id) {
         if(hash.isBlank() || hash.isEmpty()){
             try {
-                service.remove(id);
-                return ResponseEntity.status(HttpStatus.ACCEPTED).body(Boolean.TRUE);
+                UsuarioEntity user = usuarioService.findByHash(hash);
+                if(user != null){
+                    service.remove(id);
+                    return ResponseEntity.status(HttpStatus.ACCEPTED).body(Boolean.TRUE);
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
+                }
             } catch (Exception e) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Necessário enviar um identificador válido.");
             }
