+++ src/main/java/com/gabriel/martins/apinaescola/controller/EscolaController.java
@@ -1,17 +1,13 @@
 package com.gabriel.martins.apinaescola.controller;
 
 import com.gabriel.martins.apinaescola.model.entity.EscolaEntity;
+import com.gabriel.martins.apinaescola.model.entity.UsuarioEntity;
 import com.gabriel.martins.apinaescola.model.service.EscolaService;
-import java.util.Collections;
+import com.gabriel.martins.apinaescola.model.service.UsuarioService;
 import java.util.List;
-import java.util.Map;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,9 +15,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -31,6 +24,9 @@ public class EscolaController {
 
     @Autowired
     private EscolaService service;
+    
+    @Autowired
+    private UsuarioService usuarioService;
     
     @PostMapping("/escola")
     public ResponseEntity<?> cadastrarEscola(@RequestBody EscolaEntity escola){
@@ -46,8 +42,13 @@ public class EscolaController {
     public ResponseEntity<?> loginEscola(@RequestParam("hash") String hash, @RequestBody EscolaEntity escola){
         if(hash.isBlank() || hash.isEmpty()){
             try {
-                escola = service.findByLogin(escola);
-                return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
+                UsuarioEntity user = usuarioService.findByHash(hash);
+                if(user != null){
+                    escola = service.findByLogin(escola);
+                    return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
+                }
             } catch (Exception e) {
                 return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário e/ou senha incorretos.");
             }
@@ -60,8 +61,13 @@ public class EscolaController {
     public ResponseEntity<?> findById(@RequestParam("hash") String hash, @PathVariable Long id) {
         if(hash.isBlank() || hash.isEmpty()){
             try {
-                EscolaEntity escola = service.findById(id);
-                return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
+                UsuarioEntity user = usuarioService.findByHash(hash);
+                if(user != null){
+                    EscolaEntity escola = service.findById(id);
+                    return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
+                }
             } catch (Exception e) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Necessário enviar um identificador válido.");
             }
@@ -75,8 +81,13 @@ public class EscolaController {
 
         if(!hash.isBlank() || !hash.isEmpty()){
             try {
-                List<EscolaEntity> escolas = service.buscarTodos();
-                return ResponseEntity.status(HttpStatus.ACCEPTED).body(escolas);
+                UsuarioEntity user = usuarioService.findByHash(hash);
+                if(user != null){
+                    List<EscolaEntity> escolas = service.buscarTodos();
+                    return ResponseEntity.status(HttpStatus.ACCEPTED).body(escolas);
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
+                }
             } catch (Exception e) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Necessário enviar um identificador válido.");
             }
@@ -89,8 +100,13 @@ public class EscolaController {
     public ResponseEntity<?> deleteUserById(@RequestParam("hash") String hash, @PathVariable Long id) {
         if(hash.isBlank() || hash.isEmpty()){
             try {
-                service.remove(id);
-                return ResponseEntity.status(HttpStatus.ACCEPTED).body(Boolean.TRUE);
+                UsuarioEntity user = usuarioService.findByHash(hash);
+                if(user != null){
+                    service.remove(id);
+                    return ResponseEntity.status(HttpStatus.ACCEPTED).body(Boolean.TRUE);
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
+                }
             } catch (Exception e) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Necessário enviar um identificador válido.");
             }
@@ -103,8 +119,13 @@ public class EscolaController {
     public ResponseEntity<?> updateUserById(@RequestParam("hash") String hash, @PathVariable Long id, @RequestBody EscolaEntity escola) {
         if(hash.isBlank() || hash.isEmpty()){
             try {
-                service.update(escola);
-                return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
+                UsuarioEntity user = usuarioService.findByHash(hash);
+                if(user != null){
+                    service.update(escola);
+                    return ResponseEntity.status(HttpStatus.ACCEPTED).body(escola);
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Necessário autenticação.");
+                }
             } catch (Exception e) {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Requisição inválida.");
             }
